% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cortisol.R
\name{cortisol}
\alias{cortisol}
\title{Cortisol}
\usage{
cortisol(args)
}
\arguments{
\item{args}{json object. Required.}
}
\value{
json object with simulation data added.
}
\description{
Hydrocortisone Simulation API
}
\examples{
The function argument (args) must be a JSON object with the following structure:

{ 
  "args": 
   {
     "dose": 10,    // Dose (mg)
     "tint": 6,     // Interval (hurs)
     "ndos": 3,     // Number (-)
     "nseq": 1      // Period (days)
   } 
}

The function result is returned in the response body as JSON.

{ 
  "args": 
   {
     "dose": 10,    // Dose (mg)
     "tint": 6,     // Interval (hurs)
     "ndos": 3,     // Number (-)
     "nseq": 1,     // Period (days)
     "sim" :
     [
       {
         "t" :0,    // Time (h)
         "C" :0,    // Total concentration (nM)
         "Cu":0,    // Free concentration (nM)
         "fu":0     // Fraction unbound (-)
       }
     ]
   } 
}
  
DPU URLs for local single user server

Web API         : POST: http://localhost:5656/ocpu/library/cortisol/R/cortisol/json
Web APP         : GET : http://localhost:5656/ocpu/library/cortisol(/www)
Package Info    : GET : http://localhost:5656/ocpu/library/cortisol/info
R Objects       : GET : http://localhost:5656/ocpu/library/cortisol/R
R Data Objects  : GET : http://localhost:5656/ocpu/library/cortisol/data
R Help          : GET : http://localhost:5656/ocpu/library/cortisol/man/cortisol/html


}
